refactor(app): split history screen modules
